// src/proto/dapr/proto/runtime/v1/dapr.ts
export interface EncryptRequest {
  /** Request details. Must be present in the first message only. */
  options:
    | EncryptRequestOptions
    | undefined;
  /** Chunk of data of arbitrary size. */
  payload: StreamPayload | undefined;
}

/** EncryptRequestOptions contains options for the first message in the EncryptAlpha1 request. */
export interface EncryptRequestOptions {
  /** Name of the component. Required. */
  componentName: string;
  /** Name (or name/version) of the key. Required. */
  keyName: string;
  /**
   * Key wrapping algorithm to use. Required.
   * Supported options include: A256KW (alias: AES), A128CBC, A192CBC, A256CBC, RSA-OAEP-256 (alias: RSA).
   */
  keyWrapAlgorithm: string;
  /** Cipher used to encrypt data (optional): "aes-gcm" (default) or "chacha20-poly1305" */
  dataEncryptionCipher: string;
  /**
   * If true, the encrypted document does not contain a key reference.
   * In that case, calls to the Decrypt method must provide a key reference (name or name/version).
   * Defaults to false.
   */
  omitDecryptionKeyName: boolean;
  /**
   * Key reference to embed in the encrypted document (name or name/version).
   * This is helpful if the reference of the key used to decrypt the document is different from the one used to encrypt it.
   * If unset, uses the reference of the key used to encrypt the document (this is the default behavior).
   * This option is ignored if omit_decryption_key_name is true.
   */
  decryptionKeyName: string;
}

/** EncryptResponse is the response for EncryptAlpha1. */
export interface EncryptResponse {
  /** Chunk of data. */
  payload: StreamPayload | undefined;
}

/** DecryptRequest is the request for DecryptAlpha1. */
export interface DecryptRequest {
  /** Request details. Must be present in the first message only. */
  options:
    | DecryptRequestOptions
    | undefined;
  /** Chunk of data of arbitrary size. */
  payload: StreamPayload | undefined;
}

/** DecryptRequestOptions contains options for the first message in the DecryptAlpha1 request. */
export interface DecryptRequestOptions {
  /** Name of the component */
  componentName: string;
  /**
   * Name (or name/version) of the key to decrypt the message.
   * Overrides any key reference included in the message if present.
   * This is required if the message doesn't include a key reference (i.e. was created with omit_decryption_key_name set to true).
   */
  keyName: string;
}
// ...
export const EncryptRequest = {
  encode(message: EncryptRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.options !== undefined) {
      EncryptRequestOptions.encode(message.options, writer.uint32(10).fork()).ldelim();
    }
    if (message.payload !== undefined) {
      StreamPayload.encode(message.payload, writer.uint32(18).fork()).ldelim();
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): EncryptRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseEncryptRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.options = EncryptRequestOptions.decode(reader, reader.uint32());
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.payload = StreamPayload.decode(reader, reader.uint32());
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  // encodeTransform encodes a source of message objects.
  // Transform<EncryptRequest, Uint8Array>
  async *encodeTransform(
    source: AsyncIterable<EncryptRequest | EncryptRequest[]> | Iterable<EncryptRequest | EncryptRequest[]>,
  ): AsyncIterable<Uint8Array> {
    for await (const pkt of source) {
      if (globalThis.Array.isArray(pkt)) {
        for (const p of (pkt as any)) {
          yield* [EncryptRequest.encode(p).finish()];
        }
      } else {
        yield* [EncryptRequest.encode(pkt as any).finish()];
      }
    }
  },

  // decodeTransform decodes a source of encoded messages.
  // Transform<Uint8Array, EncryptRequest>
  async *decodeTransform(
    source: AsyncIterable<Uint8Array | Uint8Array[]> | Iterable<Uint8Array | Uint8Array[]>,
  ): AsyncIterable<EncryptRequest> {
    for await (const pkt of source) {
      if (globalThis.Array.isArray(pkt)) {
        for (const p of (pkt as any)) {
          yield* [EncryptRequest.decode(p)];
        }
      } else {
        yield* [EncryptRequest.decode(pkt as any)];
      }
    }
  },

  fromJSON(object: any): EncryptRequest {
    return {
      options: isSet(object.options) ? EncryptRequestOptions.fromJSON(object.options) : undefined,
      payload: isSet(object.payload) ? StreamPayload.fromJSON(object.payload) : undefined,
    };
  },

  toJSON(message: EncryptRequest): unknown {
    const obj: any = {};
    if (message.options !== undefined) {
      obj.options = EncryptRequestOptions.toJSON(message.options);
    }
    if (message.payload !== undefined) {
      obj.payload = StreamPayload.toJSON(message.payload);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<EncryptRequest>, I>>(base?: I): EncryptRequest {
    return EncryptRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<EncryptRequest>, I>>(object: I): EncryptRequest {
    const message = createBaseEncryptRequest();
    message.options = (object.options !== undefined && object.options !== null)
      ? EncryptRequestOptions.fromPartial(object.options)
      : undefined;
    message.payload = (object.payload !== undefined && object.payload !== null)
      ? StreamPayload.fromPartial(object.payload)
      : undefined;
    return message;
  },
};
// ...
export const DecryptRequestOptions = {
  encode(message: DecryptRequestOptions, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.componentName !== "") {
      writer.uint32(10).string(message.componentName);
    }
    if (message.keyName !== "") {
      writer.uint32(98).string(message.keyName);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): DecryptRequestOptions {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDecryptRequestOptions();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.componentName = reader.string();
          continue;
        case 12:
          if (tag !== 98) {
            break;
          }

          message.keyName = reader.string();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  // encodeTransform encodes a source of message objects.
  // Transform<DecryptRequestOptions, Uint8Array>
  async *encodeTransform(
    source:
      | AsyncIterable<DecryptRequestOptions | DecryptRequestOptions[]>
      | Iterable<DecryptRequestOptions | DecryptRequestOptions[]>,
  ): AsyncIterable<Uint8Array> {
    for await (const pkt of source) {
      if (globalThis.Array.isArray(pkt)) {
        for (const p of (pkt as any)) {
          yield* [DecryptRequestOptions.encode(p).finish()];
        }
      } else {
        yield* [DecryptRequestOptions.encode(pkt as any).finish()];
      }
    }
  },

  // decodeTransform decodes a source of encoded messages.
  // Transform<Uint8Array, DecryptRequestOptions>
  async *decodeTransform(
    source: AsyncIterable<Uint8Array | Uint8Array[]> | Iterable<Uint8Array | Uint8Array[]>,
  ): AsyncIterable<DecryptRequestOptions> {
    for await (const pkt of source) {
      if (globalThis.Array.isArray(pkt)) {
        for (const p of (pkt as any)) {
          yield* [DecryptRequestOptions.decode(p)];
        }
      } else {
        yield* [DecryptRequestOptions.decode(pkt as any)];
      }
    }
  },

  fromJSON(object: any): DecryptRequestOptions {
    return {
      componentName: isSet(object.componentName) ? globalThis.String(object.componentName) : "",
      keyName: isSet(object.keyName) ? globalThis.String(object.keyName) : "",
    };
  },

  toJSON(message: DecryptRequestOptions): unknown {
    const obj: any = {};
    if (message.componentName !== "") {
      obj.componentName = message.componentName;
    }
    if (message.keyName !== "") {
      obj.keyName = message.keyName;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<DecryptRequestOptions>, I>>(base?: I): DecryptRequestOptions {
    return DecryptRequestOptions.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<DecryptRequestOptions>, I>>(object: I): DecryptRequestOptions {
    const message = createBaseDecryptRequestOptions();
    message.componentName = object.componentName ?? "";
    message.keyName = object.keyName ?? "";
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends Long ? string | number | Long : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };